pyproject.toml
[project]
name = "devbase"
version = "5.1.0-alpha.4"
description = "The Definitive Engineering Operating System (EOS)"
authors = [
    { name = "Walcimar Zanque Denadai", email = "walcimarzd@gmail.com" }
]
readme = "README.md"
requires-python = ">=3.10,<3.14"
license = { text = "MIT" }
keywords = ["devops", "workspace", "cli", "johnny-decimal"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]

dependencies = [
    # CLI Framework
    "typer>=0.12.0",
    "rich>=13.7.0",
    "shellingham>=1.5.0",
    # Template Engine
    "jinja2>=3.1.6",
    "toml>=0.10.2",
    # Knowledge Management
    "networkx>=3.2",
    "python-frontmatter>=1.1.0",
]

[project.optional-dependencies]
ai = [
    # Remote LLM inference
    "groq>=1.0.0",
    # Local vector embeddings for RAG
    "fastembed>=0.7.4",
]
db = [
    # Embedded OLAP database for telemetry and analytics
    "duckdb>=0.10.0",
]
viz = [
    # Knowledge graph visualization
    "pyvis>=0.3.2",
]
dev = [
    # Project scaffolding (developer tooling only)
    "copier>=9.1.0",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]
notifications = [
    # Desktop notifications (cross-platform)
    "plyer>=2.1.0",
]
all = [
    "devbase[ai,db,viz,notifications]",
]

[project.scripts]
devbase = "devbase.main:cli_main"
db = "devbase.main:cli_main"

[project.entry-points."devbase.commands"]
core = "devbase.commands.core:app"
system = "devbase.commands.system:app"
dev = "devbase.commands.dev:app"
nav = "devbase.commands.navigation:app"
audit = "devbase.commands.audit:app"
ops = "devbase.commands.operations:app"
quick = "devbase.commands.quick:app"
docs = "devbase.commands.docs:app"
pkm = "devbase.commands.pkm:app"
study = "devbase.commands.study:app"
analytics = "devbase.commands.analytics:app"
ai = "devbase.commands.ai:app"

[project.urls]
Homepage = "https://github.com/walcimarzd/devbase-setup"
Documentation = "https://walcimarzd.github.io/devbase-setup/"
Repository = "https://github.com/walcimarzd/devbase-setup"
Changelog = "https://github.com/walcimarzd/devbase-setup/blob/main/CHANGELOG.md"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/devbase"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by formatter)
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
addopts = "--cov=devbase --cov-report=term-missing --cov-report=html --cov-fail-under=60"

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[dependency-groups]
dev = [
    "mypy>=1.14.1",
    "pytest>=8.3.5",
    "pytest-cov>=5.0.0",
    "python-semantic-release>=9.0.0",
    "ruff>=0.14.10",
]

[tool.semantic_release]
version_toml = ["pyproject.toml:project.version"]
branch = "main"
commit_message = "chore(release): {version}"
build_command = "uv build"
changelog_file = "CHANGELOG.md"
major_on_zero = false
